+++ src/components/Board/index.spec.js
@@ -109,12 +109,12 @@ describe('<Board />', () => {
             })
           })
 
-          it('calls the onCardDragEnd callback passing the card and move coordinates', () => {
+          it('calls the onCardDragEnd callback passing the card and the move coordinates', () => {
             expect(onCardDragEnd).toHaveBeenCalledTimes(1)
             expect(onCardDragEnd).toHaveBeenCalledWith(
+              { id: 1, description: 'Card content', title: 'Card title 1' },
               { fromPosition: 0, fromLaneId: 1 },
-              { toPosition: 1, toLaneId: 1 },
-              { id: 1, description: 'Card content', title: 'Card title 1' }
+              { toPosition: 1, toLaneId: 1 }
             )
           })
         })
@@ -622,7 +622,7 @@ describe('<Board />', () => {
             })
           })
 
-          it('calls the onCardDragEnd callback passing the modified board, the card and move coordinates', () => {
+          it('calls the onCardDragEnd callback passing the modified board, the card and the move coordinates', () => {
             const expectedBoard = {
               lanes: [
                 {
@@ -657,9 +657,9 @@ describe('<Board />', () => {
             expect(onCardDragEnd).toHaveBeenCalledTimes(1)
             expect(onCardDragEnd).toHaveBeenCalledWith(
               expectedBoard,
+              { description: 'Card content', id: 1, title: 'Card title 1' },
               { fromPosition: 0, fromLaneId: 1 },
-              { toPosition: 1, toLaneId: 1 },
-              { description: 'Card content', id: 1, title: 'Card title 1' }
+              { toPosition: 1, toLaneId: 1 }
             )
           })
         })
