+++ src/components/Board/services/index.js
@@ -75,40 +75,4 @@ function removeCard(board, fromLane, card) {
   return { ...board, lanes: filteredLanes }
 }
 
-function getCoordinates(event) {
-  if (event.destination === null) return {}
-
-  const laneSource = { fromPosition: event.source.index }
-  const laneDestination = { toPosition: event.destination.index }
-
-  if (isALaneMove(event.type)) {
-    return { source: laneSource, destination: laneDestination }
-  }
-
-  return {
-    source: { ...laneSource, fromLaneId: parseInt(event.source.droppableId) },
-    destination: { ...laneDestination, toLaneId: parseInt(event.destination.droppableId) }
-  }
-}
-
-function isALaneMove(type) {
-  return type === 'BOARD'
-}
-
-function getCard(board, source) {
-  const lane = board.lanes.find(lane => lane.id === source.fromLaneId)
-  return lane.cards[source.fromPosition]
-}
-
-export {
-  moveLane,
-  moveCard,
-  addLane,
-  removeLane,
-  renameLane,
-  addCard,
-  removeCard,
-  getCard,
-  getCoordinates,
-  isALaneMove
-}
+export { moveLane, moveCard, addLane, removeLane, renameLane, addCard, removeCard }
