+++ src/components/Board/services/index.spec.js
@@ -1,15 +1,4 @@
-import {
-  moveLane,
-  moveCard,
-  addLane,
-  removeLane,
-  renameLane,
-  addCard,
-  removeCard,
-  getCard,
-  isALaneMove,
-  getCoordinates
-} from './'
+import { moveLane, moveCard, addLane, removeLane, renameLane, addCard, removeCard } from './'
 
 describe('#moveLane', () => {
   it('returns a board with the lane moved to the specified position', () => {
@@ -197,64 +186,5 @@ describe('#removeCard', () => {
         { id: 2, cards: [{ id: 4 }, { id: 5 }, { id: 6 }] }
       ]
     })
-  })
-})
-
-describe('#getCoordinates', () => {
-  describe('when the event does not have destination', () => {
-    it('returns empty object', () => {
-      expect(getCoordinates({ destination: null })).toEqual({})
-    })
-  })
-
-  describe('when the event is a lane move', () => {
-    const event = {
-      type: 'BOARD',
-      source: { index: 0, droppableId: 0 },
-      destination: { index: 1, droppableId: 1 }
-    }
-
-    it('returns a object with the lane coordinates', () => {
-      expect(getCoordinates(event)).toEqual({ destination: { toPosition: 1 }, source: { fromPosition: 0 } })
-    })
-  })
-
-  describe('when the event is not a lane move', () => {
-    const event = {
-      type: 'CARD',
-      source: { index: 0, droppableId: 0 },
-      destination: { index: 1, droppableId: 1 }
-    }
-
-    it('returns a object with the card coordinates', () => {
-      expect(getCoordinates(event)).toEqual({
-        destination: { toLaneId: 1, toPosition: 1 },
-        source: { fromLaneId: 0, fromPosition: 0 }
-      })
-    })
-  })
-})
-
-describe('#isALaneMove', () => {
-  describe('when the type is "BOARD"', () => {
-    it('returns true', () => {
-      expect(isALaneMove('BOARD')).toEqual(true)
-    })
-  })
-
-  describe('when the type is not "BOARD"', () => {
-    it('returns false', () => {
-      expect(isALaneMove('IS_NOT_BOARD')).toEqual(false)
-    })
-  })
-})
-
-describe('#getCard', () => {
-  const board = {
-    lanes: [{ id: 1, cards: [{ id: 1 }, { id: 2 }, { id: 3 }] }]
-  }
-
-  it('returns the card of the board from the given source', () => {
-    expect(getCard(board, { fromLaneId: 1, fromPosition: 1 })).toEqual({ id: 2 })
   })
 })
